requirements.txt
asgiref==3.7.2 
async-timeout==4.0.3
attrs==23.1.0 
autobahn==23.6.2 
automat==22.10.0 
certifi==2023.7.22 
cffi==1.16.0 
channels-redis==4.1.0 
channels==4.0.0 
channels[daphne]==4.0.0 
charset-normalizer==3.3.2 
click==8.1.7 
colorama==0.4.6
constantly==23.10.4 
cryptography==41.0.5 
daphne==4.0.0 
defusedxml==0.7.1 
django-allauth-ui==0.1.5 
django-allauth==0.58.2 
django-widget-tweaks==1.5.0 
django==4.2.7 
exceptiongroup==1.2.0 
h11==0.14.0 
h2==4.1.0 
hpack==4.0.0 
hypercorn==0.15.0 
hyperframe==6.0.1 
hyperlink==21.0.0 
idna==3.4 
incremental==22.10.0 
msgpack==1.0.7 
mysqlclient==2.2.0 
oauthlib==3.2.2 
priority==2.0.0 
psycopg-binary==3.1.13 
psycopg[binary]==3.1.13 
pyasn1-modules==0.3.0 
pyasn1==0.5.0 
pycparser==2.21 
pyjwt[crypto]==2.8.0 
pyopenssl==23.3.0 
python-dotenv==1.0.0 
python3-openid==3.2.0 
redis==5.0.1 
requests-oauthlib==1.3.1 
requests==2.31.0 
service-identity==23.1.0 
setuptools==68.2.2 
six==1.16.0 
sqlparse==0.4.4 
taskgroup==0.0.0a4 
tomli==2.0.1 
twisted[tls]==23.10.0 
txaio==23.1.1 
typing-extensions==4.8.0 
tzdata==2023.3
urllib3==2.1.0 
uvicorn==0.24.0.post1 
whitenoise==6.6.0 
wsproto==1.2.0 
zope-interface==6.1
google-generativeai==0.3.2
django-ckeditor==6.7.0
openai==1.9.0